Fix PrimitiveCache flush to accept kind and use app container in terminating callback
